Support unsuccessful request responses
When a TSP request response is received with an unsuccessful status code
and without a json object model, the parsing of the body as json was
throwing an error and the status code was unavailable to the caller.

Create the TspClientResponse with the plain text of the response, and
make the model optional. Make the plain text available to the caller. Do
not throw an error when the json object is absent.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>