Serialize and deserialize timestamps as strings
Fields that are known to be timestamps or durations are serialized and
deserialized as strings.

This is done to avoid loss of precision in JavaScript 'number' type for
values above Number.MAX_SAFE_INTEGER (2^53 - 1).

The models are updated to use 'bigint' type instead of 'number' for
these fields.

In outgoing requests, bigint values are replaced by strings in the
JSON.stringify() replacer. The strings are then replaced by unquoted
numbers in the JSON request.

In incoming JSON responses, long numbers are replaced by strings prior
to parsing, and the JSON.parse() reviver will replace those strings by
either bigint or number values, depending on the field. For bigint
fields, small numbers are replaced by their bigint value.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>