depend on json-bigint from the right package
What happend is that the PR that introduced the dependency was done
before a refactoring that moved the `tsp-typescript-client` package in a
different folder. But the change to the root `package.json` dependencies
didn't cause a merge conflict, so it stayed this way even though it
meant that the `tsp-typescript-client` would be published without the
new dependency.

Move the dependency from the root of the monorepo to the
`tsp-typescript-client` package.

Signed-off-by: Paul Marechal <paul.marechal@ericsson.com>