Add ConnectionStatusListener to RestClient
Clients can add a connection status listener to the RestClient to be
notified when the connection goes up or down. The connection status is
detected based on whether any HTTP request can be completed or not, or
based on the returned HTTP status code.

When a listener is added it will immediately receive the current status
through the listener interface.

Change-Id: I5f7fcbf4b94380b2ea349755a400bd61bbadc7d2
Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>