Make TspClient reuseable interface
The new interface is called `ITspClient`.
`ITspClient` can be used for typing and other implementations.
The existing functionality was moved to the `HttpTspClient`.
To not break the existing behavior `TspClient` is still exported as an
alias for the `HttpTspClient`, but is being deprecated.
This way adopters should be notified to migrate to `HttpTspClient`.
This makes the package more flexible as different implementations
can be provided, e.g. a proxy client in a Theia frontend.

Part of https://github.com/eclipse-cdt-cloud/theia-trace-extension/issues/976

Contributed on behalf of STMicroelectronics