[lerna] upgrade to recent version
During the "yarn upgrade" that came along the lerna upgrade, it was
found that the "tslint" dependencies were no longer valid (deprecated,
old).

Manually running tslint from the master branch revealed that the
repo's code has not been following the configured rules, probably
for a long time (>100 violations). In consequence, I think the best
approach at this time is to remove tslint's dependencies and configs.

We can consider adding ESLint to the repo, separately, if desired.

Fixes #84

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>