Align OutputConfigurationQuery/Configuration for sourceTypeId
In the OutputConfigurationQuery the configuration source type id was
called typeId and it creates a Configuration object where the same field
is called sourceTypeId for the same field. When serializing the
Configuration in the derived DataProviderDescriptor over TSP the field
is called sourceTypeId and this is not consistent. The returned
Configuration should have the exact field names that are used when
creating the data provider and the respective configuration.

The Configuration is already API while OutputConfigurationQuery is not,
and it's better to change the name there.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>