Further fix parsing of sampling to match tsp specification
The server was modified to comply to tsp specification for
sampling, update accordingly here. There are now 3 different arrays
for x values: xValues (legacy xy for timestamps), xCategories (array of
string) and xRanges (array of StartEndRanges).

Note that StartEndRange is now an object with start and end attributes.
Before it was an array with 2 values.

Contributes to fix:
https://github.com/eclipse-tracecompass-incubator/org.eclipse.tracecompass.incubator/issues/236

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>